server/requirements.txt

asgiref==3.8.1
Django==4.2.8
django-otp==1.5.4
django-use-email-as-username==1.4.0
pyotp==2.9.0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
